requirements.txt
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.2.17
django-allauth==0.50.0
django-cors-headers==4.7.0
django-filter==24.3
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
oauthlib==3.2.2
pillow==10.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==75.8.2
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
zope.interface==7.2
tldextract